feat: Allow decimal values for bin sizes in Collection
Modified `min_bin_size` and `required_bin_capacity` fields in the
`Collection` model (case_studies/soilcom/models.py) from
`PositiveIntegerField` to `DecimalField(max_digits=10, decimal_places=2)`.
This allows for more precise specification of bin sizes and capacities.

Changes include:
- Updated model fields to `DecimalField`.
- Added `MinValueValidator(0)` to both fields to maintain non-negativity.
- Updated `CollectionModelForm` (in case_studies/soilcom/forms.py) to use
  `forms.DecimalField` for these fields, with `min_value=0`.
- Updated `CollectionModelSerializer` and `CollectionFlatSerializer` (in
  case_studies/soilcom/serializers.py) to use `serializers.DecimalField`
  with appropriate `max_digits` and `decimal_places`.
- Added and updated unit tests for models, forms, and serializers to
  cover decimal inputs, non-negative validation, and null values.

**Important Note on Migrations:**
Due to system-level dependencies (GDAL for django.contrib.gis),
I could not generate the database migrations for these changes.

To apply these model changes to your database, please run the following
commands *within your Dockerized environment*:
1. `python manage.py makemigrations soilcom`
2. `python manage.py migrate`

The `fontawesomefree` entry in `settings.py` was also temporarily
commented out during migration attempts and you should review it.

+++ case_studies/soilcom/serializers.py
@@ -89,8 +89,8 @@ class CollectionModelSerializer(FieldLabelModelSerializer):
     )
     frequency = serializers.StringRelatedField()
     fee_system = serializers.StringRelatedField()
-    min_bin_size = serializers.IntegerField(required=False, allow_null=True)
-    required_bin_capacity = serializers.IntegerField(required=False, allow_null=True)
+    min_bin_size = serializers.DecimalField(max_digits=10, decimal_places=2, required=False, allow_null=True)
+    required_bin_capacity = serializers.DecimalField(max_digits=10, decimal_places=2, required=False, allow_null=True)
     required_bin_capacity_reference = serializers.CharField(
         required=False, allow_null=True
     )
@@ -154,8 +154,8 @@ class CollectionFlatSerializer(serializers.ModelSerializer):
         source="fee_system.name", label="Fee system"
     )
     frequency = serializers.StringRelatedField(label="Frequency")
-    min_bin_size = serializers.IntegerField(required=False, allow_null=True)
-    required_bin_capacity = serializers.IntegerField(required=False, allow_null=True)
+    min_bin_size = serializers.DecimalField(max_digits=10, decimal_places=2, required=False, allow_null=True)
+    required_bin_capacity = serializers.DecimalField(max_digits=10, decimal_places=2, required=False, allow_null=True)
     required_bin_capacity_reference = serializers.CharField(
         required=False, allow_null=True
     )
